Fix integer overflow at the reservoir sampling
As the Reservoir sampling implementation is shared across shards and
partitions, it may process more documents than an integer value can hold.
It must be protected against integer overflows to not add more that
the defined maximum samples to the list (this limit is also in the
integer range and such supporting sample sizes > integer is not needed).

The integer overflow results in bypassing the maximum samples boundary
and can lead to node crashes due to OutOfMemory exceptions.